[#42]refactoring: user 폴더 추가
