+++ src/main/java/com/runninghi/user/command/application/service/UserService.java
@@ -1,9 +1,9 @@
 package com.runninghi.user.command.application.service;
 
-import com.runninghi.user.command.application.dto.request.UserUpdateRequest;
-import com.runninghi.user.command.application.dto.response.UserDeleteResponse;
-import com.runninghi.user.command.application.dto.response.UserInfoResponse;
-import com.runninghi.user.command.application.dto.response.UserUpdateResponse;
+import com.runninghi.user.command.application.dto.user.request.UserUpdateRequest;
+import com.runninghi.user.command.application.dto.user.response.UserDeleteResponse;
+import com.runninghi.user.command.application.dto.user.response.UserInfoResponse;
+import com.runninghi.user.command.application.dto.user.response.UserUpdateResponse;
 import com.runninghi.user.command.domain.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
