+++ src/main/java/com/runninghi/user/command/domain/aggregate/entity/User.java
@@ -1,8 +1,8 @@
 package com.runninghi.user.command.domain.aggregate.entity;
 
 import com.runninghi.common.entity.BaseEntity;
-import com.runninghi.user.command.application.dto.request.UserUpdateRequest;
 import com.runninghi.user.command.application.dto.sign_up.request.SignUpRequest;
+import com.runninghi.user.command.application.dto.user.request.UserUpdateRequest;
 import com.runninghi.user.command.domain.aggregate.entity.enumtype.Role;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
